Refactor: Build.gradle
