+++ app/src/main/java/org/retriever/dailypet/interfaces/MyFirebaseMessagingService.kt
@@ -5,16 +5,15 @@ import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import android.content.SharedPreferences
 import android.media.RingtoneManager
 import android.os.Build
 import android.util.Log
 import androidx.core.app.NotificationCompat
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
+import org.retriever.dailypet.GlobalApplication
 import org.retriever.dailypet.MainActivity
 import org.retriever.dailypet.R
-import org.retriever.dailypet.models.App
 
 class MyFirebaseMessagingService : FirebaseMessagingService() {
 
@@ -24,7 +23,7 @@ class MyFirebaseMessagingService : FirebaseMessagingService() {
     override fun onNewToken(token: String) {
         Log.d(TAG, "new Token: $token")
         // 토큰 값을 따로 저장해둔다.
-        App.prefs.setString("deviceToken",token)
+        GlobalApplication.prefs.setString("deviceToken",token)
         Log.i("로그: ", "성공적으로 토큰을 저장함")
     }
