+++ build.gradle
@@ -1,11 +1,15 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     repositories{
+        google()
+        mavenCentral()
+
         maven { url 'https://devrepo.kakao.com/nexus/content/groups/public/'}
     }
     dependencies {
-        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.9.1'
-        classpath 'com.google.gms:google-services:4.3.13'
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.9.2'
+        classpath 'com.google.gms:google-services:4.3.14'
+        classpath 'androidx.navigation:navigation-safe-args-gradle-plugin:2.5.2'
     }
 }
