+++ gradle.properties
@@ -21,3 +21,9 @@ kotlin.code.style=official
 # resources declared in the library itself and none from the library's dependencies,
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
+
+BASE_URL="http://ec2-13-124-5-128.ap-northeast-2.compute.amazonaws.com:8080/"
+KAKAO_NATIVE_APP_KEY="45276ed14e089e697b39bb9cab043b7f"
+MANIFEST_KAKAO_NATIVE_APP_KEY="kakao45276ed14e089e697b39bb9cab043b7f"
+NAVER_CLIENT_ID="bl7aZ1ovzbnYvXH8ERnu"
+NAVER_CLIENT_SECRETE="3WB1kGyF_R"
